requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-SocketIO==5.3.6
Flask-Migrate==4.0.5
gunicorn==21.2.0
eventlet==0.33.3
python-socketio==5.8.0
python-engineio==4.7.1
Werkzeug==2.3.7
SQLAlchemy==2.0.21
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3
dnspython==2.4.2
six==1.16.0
greenlet==3.0.1
psutil==5.9.6
APScheduler==3.10.4
python-dotenv==1.0.0
